Some initial work on GOAT pulse optimization
After reading the paper, it seems like the best place to implement this is in XACC, not QuaC since this has nothing to do with state evolution, which is what QuaC does. Perhaps, we just need to verify the final *optimized* pulse on QuaC.

This commit contains high-level (abstract) components to implement GOAT. Due to its analytical nature, I think it's a good idea to provide an abstract way to specify the algorithm, e.g. we could implement problem-specific hooks if needed.

Signed-off-by: Thien Nguyen <nguyentm@ornl.gov>