Working example of GOAT
This is still *Work-in-progress* at the moment.
In particular, I'm using a simple L-BFGS lib for testing purposes.
Interestingly enough, both NL-OPT and ML-PACK failed to optimize (using l-bfgs)
this example. They were simply stuck. All impls followed the same trajectory but only
the simple one was able to escape the local minima and hence converge.

Want to get this checked-in as the baseline first while diving into why ML-PACK and NL-OPT
cannot work out-of-box for this case. Perhaps there are some options that we can tweak.

Tested by: the added unit test for simple X gate optimization

Signed-off-by: Thien Nguyen <nguyentm@ornl.gov>