Fixed exp-val-z calculation for the Qpp Accelerator
Swapping the order of conditional check so that in we can further optimize for the case where there is no If statements in the circuit nor shots; in that case, calculate expectation analytically.

Also, fixed a potential bug whereby the shots count may not be reset if using the accelerator multiple times.

Signed-off-by: Thien Nguyen <nguyentm@ornl.gov>