Potential fix for the *false positive* explore phase
The explore phase (locations as parameters) using soft-max to localize floating-point parameters.

This worked in most of the cases unless the optimizer *accidentally* picked two identical location params at one layer and it created a local minimum.

This can potentially terminate the explore phase early with less layers but some layers are combinations of multiple locations instead of just one location.

Hence, the refine stage cannot find the solution when it fixes the location.

Signed-off-by: Thien Nguyen <nguyentm@ornl.gov>