Initial Thread-Safe Implementation
Add thread-safety to XACC/QCOR (Experimental). We introduce `_XACC_MUTEX` macro in `xacc.hpp` to turn on/off the feature (default is OFF). When the macro is ON, only a single thread can execute the following routines by using either `std::mutex` or `std::recursive_mutex`:

- User-facing API routines
   - `xacc::qalloc()`

- Backend modules
   - `QppAccelerator`
       - `execute()` (the existing shot-level parallel execution still works)
       - `apply()`
   - `NLOptimizer`
      -  `optimize()`

Also, regardless of the macro, the `NLOptimizer` class is now `xacc::Cloneable`.

Signed-off-by: Akihiro Hayashi <ahayashi@gatech.edu>