IBM: Make bidirectional couplings when server sends no coupling_map
In the IBM backend, we require [x, y] to exist in the coupling map to
execute CX x,y -- [y, x] in the coupling map is not enough. So when IBM
does not include a coupling_map for a given backend, we need to include
both [i,j] and [j,i] for all qubits i,j s.t. i != j in the coupling map
we create; currently, we only include [i,j] for all qubits i,j with i <
j, which will cause errors if a circuit includes CX j,i.

Signed-off-by: Austin Adams <aja@gatech.edu>