Fixed an issue with MCX gen using the gray code method
The problem is that we are using Rx(pi) == X, ignoring the global phase value of -i.
We need to use H-MCZ-H to construct MCX.

Note: the existing MCX test expected 220 U3 gates (from the code comment), which is now true thanks to this fix. Also, added a test to validate the matrix and truth table.

Also, fixed https://github.com/eclipse/xacc/issues/514

Signed-off-by: Thien Nguyen <thien.md.nguyen@gmail.com>