Fixed qpp C-U simulation
The original impl tried to be generic using the GateFuser util. However, this created a problem when the base U is not on the LSB qubits. GateFuser will generate kron matrix causing dimension mismatch in qpp code.

Hence, just limit the use to the common X, Y, Z case -> no need for GateFuser.

Signed-off-by: Thien Nguyen <thien.md.nguyen@gmail.com>