Improve to C-U gate application
For C-U, we currently have a fast path for gray-code decomposition for a limited set of gates.
The issue is when the base circuit contains already controlled gates, e.g., CNOT, CZ, etc., we cannot handle this case efficiently.

In this CS, we added the ability to recognize these control-like instructions and combine control qubits when C-U is applied on them.

For example, C-U(0,1,2,3) on CNOT(4, 5) ==> C-U(0,1,2,3,4) on X(5), etc.

Signed-off-by: Thien Nguyen <thien.md.nguyen@gmail.com>